move model interfaces to model/~
