--- src/data/Card.tsx
@@ -1,8 +1,0 @@
-import { ViewState } from 'src/cards/CardRenderer'
-
-export interface Card {
-  deckID: number
-  cardTypeID: number
-  fields: { [K in ViewState]: React.ReactNode }
-  tags: Array<string>
-}
