+++ src/model/interfaces.tsx
@@ -1,0 +1,5 @@
+export { Card } from 'src/model/Card'
+export enum ViewState {
+  Front = 'front',
+  Back = 'back',
+}
